+++ demo/js/src/ie/matchMedia.ie8.js
@@ -1,3 +1,9 @@
+// Modernizr style test
+if (!(window.webkitMatchMedia || window.mozMatchMedia || window.oMatchMedia || window.msMatchMedia || window.matchMedia)) {
+	var root = document.getElementsByTagName( 'html' )[0];
+	root.className += ' no-matchmedia';
+}
+
 /*! matchMedia() polyfill - Test a CSS media type/query in JS. Authors & copyright (c) 2012: Scott Jehl, Paul Irish, Nicholas Zakas. Dual MIT/BSD license - IE8 VERSION! */
 window.matchMedia = window.matchMedia || (function(doc, undefined){